requirements.txt
-e .
aws-cdk.assets
aws-cdk.aws-apigateway
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling
aws-cdk.aws-autoscaling-common
aws-cdk.aws-autoscaling-hooktargets
aws-cdk.aws-cloudwatch
aws-cdk.aws-ec2
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-eks
aws-cdk.aws-elasticloadbalancing
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.region-info